refactor: split Planet aggregate into partial classes by concern (#40)
Move-only: Planet.cs keeps state + RebaseRates/CheckpointAllResources;
energy, building lifecycle, and ship queue/roster move to
Planet.Energy.cs / Planet.Buildings.cs / Planet.Ships.cs. Marten still
sees one type. Size/responsibility guideline + audit recorded in
project-structure.md; suite green with zero test edits.
